rename func tests
