+++ func_tests/test_client.py
@@ -1,4 +1,4 @@
-from func_test.vinyldns_context import *
+from func_tests.vinyldns_context import vinyldns_test_context
 
 def test_list_zones(vinyldns_test_context):
     zones = vinyldns_test_context.client.list_zones()
